[Modify] post cbv view => function view

+++ mysite/blog/views.py
@@ -4,38 +4,36 @@ from blog.models import Post
 from blog.serializers import PostSerializer
 from rest_framework.response import Response
 from rest_framework import status
+from rest_framework.decorators import api_view
 
 
-class PostListview(APIView):
-    def get(self, request):
-        qs = Post.objects.all()
-        serializer = PostSerializer(qs, many=True)
+@api_view(['GET', 'POST'])
+def post(request):
+    if request.method == 'GET':
+        serializer = PostSerializer(Post.objects.all(), many=True)
         return Response(serializer.data)
-    def post(self, request, format=None):
+    else:
         serializer = PostSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=201)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(serializer.data, status=400)
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
-class PostDetailView(APIView):
-    def get_object(self, pk):
-        return get_object_or_404(Post, pk=pk)
-    def get(self, request, pk, format=None):
-        post = self.get_object(pk)
+@api_view(['GET', 'PUT', 'DELETE'])
+def post_detail(request, pk):
+    post = get_object_or_404(Post, pk=pk)
+    if request.method == 'GET':
         serializer = PostSerializer(post)
         return Response(serializer.data)
-    def put(self, request, pk, format=None):
-        post = self.get_object(pk)
+    elif request.method == 'PUT':
         serializer = PostSerializer(post, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         else:
             return Response(serializer.error, status=status.HTTP_400_BAD_REQUEST)
-    def delete(self, request, pk, format=None):
-        post = self.get_object(pk)
+    else:
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
